Fix: responsiveness and content for index and tour-page
Index page: refactored some of the code in the hero-section to look more cohesive with rest of page
and for better responsiveness.
Tour page: Changed styling for tour schedule and changed it from using flex to grid
for better responsiveness.

Changed some styling for "components" such as buttons and email-fields.